Migrator tests: fixes & phpstan

+++ tests/MigratorTest.php
@@ -17,9 +17,18 @@ use function PHPUnit\Framework\assertCount;
 
 final class MigratorTest extends TestCase
 {
+    private Connection $connection;
+
     public function __construct(string $name)
     {
-        $config = json_decode(file_get_contents(__DIR__ . '/../vendor/doomy/testing/testingDbCredentials.json'), true);
+        $rawConfig = file_get_contents(__DIR__ . '/../vendor/doomy/testing/testingDbCredentials.json');
+        if ($rawConfig === false) {
+            throw new \Exception('Could not read config');
+        }
+        $config = json_decode($rawConfig, true);
+        if (! is_array($config)) {
+            throw new \Exception('Invalid config');
+        }
         $this->connection = new Connection($config);
         parent::__construct($name);
     }
@@ -29,28 +38,29 @@ final class MigratorTest extends TestCase
         $this->connection->query('DROP TABLE t_migration');
         $this->connection->query('DROP TABLE if exists t_test');
 
-        foreach (glob(__DIR__ . '/migrations' . '/*') as $file) {
-            is_file($file) && unlink($file);
+        $migrationFiles = glob(__DIR__ . '/migrations/*');
+        if (is_array($migrationFiles)) {
+            foreach ($migrationFiles as $file) {
+                is_file($file) && unlink($file);
+            }
         }
     }
 
     public function testMigrateCreatesTable(): void
     {
-        $config = json_decode(file_get_contents(__DIR__ . '/../vendor/doomy/testing/testingDbCredentials.json'), true);
-        $connection = new Connection($config);
-        $entityFactory = new EntityFactory($connection);
-        $repoFactory = new RepoFactory($connection, $entityFactory);
+        $entityFactory = new EntityFactory($this->connection);
+        $repoFactory = new RepoFactory($this->connection, $entityFactory);
         $data = new DataEntityManager($repoFactory, new EntityCache());
-        $migrator = new Migrator($connection, $data, [
+        $migrator = new Migrator($this->connection, $data, [
             'migrations_directory' => __DIR__ . '/migrations',
         ]);
 
-        $tables = $connection->query('SHOW TABLES')
+        $tables = $this->connection->query('SHOW TABLES')
             ->fetchAll();
         Assert::assertCount(0, $tables);
 
         $migrator->migrate();
-        $tables = $connection->query('SHOW TABLES')
+        $tables = $this->connection->query('SHOW TABLES')
             ->fetchAll();
         Assert::assertCount(1, $tables);
         Assert::assertEquals('t_migration', $tables[0]['Tables_in_testing']);
@@ -58,12 +68,10 @@ final class MigratorTest extends TestCase
 
     public function testMigrateAppliesMigration(): void
     {
-        $config = json_decode(file_get_contents(__DIR__ . '/../vendor/doomy/testing/testingDbCredentials.json'), true);
-        $connection = new Connection($config);
-        $entityFactory = new EntityFactory($connection);
-        $repoFactory = new RepoFactory($connection, $entityFactory);
+        $entityFactory = new EntityFactory($this->connection);
+        $repoFactory = new RepoFactory($this->connection, $entityFactory);
         $data = new DataEntityManager($repoFactory, new EntityCache());
-        $migrator = new Migrator($connection, $data, [
+        $migrator = new Migrator($this->connection, $data, [
             'migrations_directory' => __DIR__ . '/migrations',
         ]);
         $migrationFilename = '01-testing-migration.sql';
@@ -72,7 +80,7 @@ final class MigratorTest extends TestCase
             'CREATE TABLE t_test (id INT NOT NULL AUTO_INCREMENT, PRIMARY KEY(id));'
         );
         $migrator->migrate();
-        $tables = $connection->query('SHOW TABLES')
+        $tables = $this->connection->query('SHOW TABLES')
             ->fetchAll();
         Assert::assertCount(2, $tables);
         Assert::assertEquals('t_test', $tables[1]['Tables_in_testing']);
